Clear the clang-cl build of every error and warning
- Handle the missing FileKind::Shader and PixelArt tool cases the switch-enum
  errors were pointing at; a dropped shader file had been falling through to the
  generic "file" label
- Spell out TOML's date/time node types instead of letting them default into the
  scene cache as Null
- Take the Vulkan image-layout lookup off switch-enum: it is an open vendor enum
  that grows with every SDK, not an exhaustive dispatch
- Drop dead locals, captures, fields and a duplicated shader constant
- Guard the dotnet-env and runtime-name helpers with the same condition as their
  only call sites
- Put RenderGraph's move constructor in declaration order
- Read the temp dir through std::filesystem rather than a deprecated getenv
- Turn the unused physics layer count into a static_assert that keeps it honest

+++ src/app/debug/PerformancePanel.cpp
@@ -82,7 +82,7 @@ namespace aether::editor
 			ImGui::TableSetColumnIndex(0);
 			ImGui::TextUnformatted("Frame");
 			ImGui::SameLine();
-			ImGui::Text("#%u", context.frameIndex);
+			ImGui::Text("#%llu", static_cast<unsigned long long>(context.frameIndex));
 
 			ImGui::TableSetColumnIndex(1);
 			ImGui::TextUnformatted("FPS ");
